Completed.

+++ StringSet.cpp
@@ -46,7 +46,7 @@ namespace sdds {
 	{
 		ss_string_count = input_string.ss_string_count;
 		ss_strings = new string[ss_string_count];
-		for (int i = 0; i < ss_string_count; ++i)
+		for (size_t i = 0; i < ss_string_count; ++i)
 		{
 			ss_strings[i] = input_string.ss_strings[i];
 		}
